Move feed, reputation, webhook delivery under services/
Implementations live in services/{feed_impl,reputation,webhook_delivery}.
Keep app-level shims for historical imports.
