+++ chafan_core/app/api/api_v1/endpoints/admin_tools.py
@@ -4,7 +4,7 @@ from fastapi import APIRouter, Depends, Response
 from chafan_core.app.config import settings
 from chafan_core.app.api import deps
 from chafan_core.app.infra.request_context import RequestContext
-from chafan_core.app.feed import get_site_activities
+from chafan_core.app.services.feed_impl import get_site_activities
 from chafan_core.utils.base import HTTPException_
 from chafan_core.app.responders.rss import build_rss
